Create UCL_DOAS_readMiniMaxLog.py
MiniMax logs record all operations by the MAX-DOAS. Of particular interest are the automatic azimuth direction finding function which update the set azimuth north position of the instrument as a result of looking for the sun. The elevation of the sun found relative to that calculated is also recorded.
This code reads the MiniMax log and plots the azimuth north position and the difference in calculated vs measured solar position.